+++ coverage/ctracer/tracer.c
@@ -30,6 +30,11 @@ static PyObject *str__coverage_plugin_name;
 static PyObject *str_dynamic_source_filename;
 static PyObject *str_line_number_range;
 
+#if PY_VERSION_HEX >= 0x030D0000
+/* Non-static: declared in util.h for the MyFrame_SetTrace macro. */
+PyObject *str_f_trace;
+#endif
+
 int
 CTracer_intern_strings(void)
 {
@@ -45,6 +50,9 @@ CTracer_intern_strings(void)
     INTERN_STRING(str__coverage_plugin_name, "_coverage_plugin_name")
     INTERN_STRING(str_dynamic_source_filename, "dynamic_source_filename")
     INTERN_STRING(str_line_number_range, "line_number_range")
+#if PY_VERSION_HEX >= 0x030D0000
+    INTERN_STRING(str_f_trace, "f_trace")
+#endif
 
     ret = RET_OK;